reuse the existing truncation test
